requirements.txt
slack_bolt
aiohttp
python-decouple
fastapi<1
uvicorn<1
uvloop
httptools
requests
pytz
oauthlib
requests_oauthlib